Latest Commit
- Levenshtein #5
- Added test txt files
- Added crypt #1

+++ program-filer/helpers.h
@@ -1,8 +1,11 @@
+#include <stdio.h>
+#include <string.h>
 
 /* print char array as string */
 void print_str(char array[]) {
-    for (int i = 0; i < strlen(array); i++) {
+    for (unsigned long i = 0LU; i < strlen(array); i++) {
         printf("%c", array[i]);
     }
     printf("\n");
 }
+
